requirements.txt
amqp==2.6.1
appdirs==1.4.4
apt-xapian-index==0.49
asn1crypto==1.4.0
billiard==3.6.4.0
blinker==1.4
CacheControl==0.12.10
cachetools==4.2.4
cachy==0.3.0
certifi==2020.6.20
chardet==4.0.0
cleo==0.8.1
click==7.1.2
clikit==0.6.2
colorama==0.4.4
command-not-found==0.3
crashtest==0.3.1
croniter==1.3.4
cryptography==3.3.2
cupshelpers==1.0
dateutils==0.6.12
dbus-python==1.2.16
defer==1.0.6
Deprecated==1.2.13
distlib==0.3.1
distro==1.5.0
distro-info==1.0
django-braces==1.14.0
django-discover-runner==1.0
docopt==0.6.2
elasticsearch==7.17.1
et-xmlfile==1.1.0
filelock==3.0.12
google-api-core==1.31.5
google-api-python-client==2.40.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.3
google-cloud-firestore==2.4.0
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.55.0
gpg===1.14.0-unknown
grpcio==1.44.0
html5lib==1.1
httplib2==0.18.1
idna==2.10
importlib-metadata==1.6.0
jdatetime @ git+https://github.com/Meysam-/python-jalali@1bf2b2df6e76925db3ecf1f5069458a06f8c5550
jdcal==1.4.1
jeepney==0.6.0
Jinja2==3.1.1
keyring==22.2.0
kombu==4.6.11
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lockfile==0.12.2
macaroonbakery==1.3.1
MarkupSafe==2.1.1
mechanize==0.4.5
more-itertools==4.2.0
msgpack==1.0.3
netifaces==0.10.9
oauthlib==2.0.1
olefile==0.46
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
Pillow==8.1.2
pkginfo==1.8.2
poetry==1.1.13
poetry-core==1.0.8
proto-plus==1.20.3
protobuf==3.19.4
protonvpn-cli==2.2.11
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycups==2.0.1
PyGObject==3.38.0
PyJWT==1.7.1
pylev==1.4.0
pymacaroons==0.13.0
pymongo==4.0.2
PyNaCl==1.4.0
pyparsing==3.0.7
PyQt5==5.15.4
PyQt5-sip==12.8.1
pyRFC3339==1.1
python-apt===2.2.0-ubuntu0.21.04.1
python-debian==0.1.39
pythondialog==3.5.3
pytz==2021.3
PyYAML==5.3.1
redis==4.1.4
reportlab==3.5.66
requests==2.25.1
requests-toolbelt==0.9.1
rq==1.10.1
rsa==4.8
SecretStorage==3.3.1
shellingham==1.4.0
simplejson==3.17.2
six==1.15.0
systemd-python==234
tomlkit==0.10.2
ubuntu-advantage-tools==27.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.2
usb-creator==0.3.7
vine==1.3.0
virtualenv==20.4.0+ds
wadllib==1.3.5
webencodings==0.5.1
wrapt==1.14.0
xkit==0.0.0
zipp==1.0.0
